fix(stability): close 3 UBSan findings surfaced by ASan/UBSan run
A clean-build run with -fsanitize=address,undefined surfaced three
distinct undefined-behaviour sites under the existing test suite +
the new gates.  All three are pre-existing — not introduced by any
recent change — but they're real correctness risks on strict-
alignment ABIs and, for the int64 overflow, on any future compiler
that decides to optimise on the assumption.  Fixed in place; ASan/
UBSan re-run is now clean (0 SUMMARY: lines across 56+2 binaries).

1. src/compress/pfor.c:380  signed int64 overflow

   The adversarial pfor encoder test feeds the loop ts pairs near the
   signed-range boundary (e.g. +7.2e18 vs -6.6e18).  Plain
   `in[i+k] - prev` overflowed int64.  Two's-complement wrap is what
   we want anyway — same bit pattern reaches the zigzag encode — so
   route the subtraction through uint64_t to make the wrap defined.

2. src/storage/memtable.c:487  pointer arithmetic on misaligned
   const int64_t *
3. src/storage/memtable.c:507  load of misaligned int64 via array
   subscript

   Both fire from the WRITE_BATCH RPC receive path (rpc.c:386) where
   ts_arr is a slice into a wire buffer at a non-8-aligned offset.
   The C standard requires int64_t* to be properly aligned even before
   dereference; cast at the function-entry was UB.  Fix: change both
   tsdb_memtable_append_bulk and tsdb_batch_append_bulk to take
   `const void *ts_arr`, memcpy out 8 bytes per row inside.  Six
   callsites in db.c / memtable.c / influx_line.c / server.c / rpc.c
   compile clean against the new signatures (the int64_t* → void*
   conversion at the call site is implicit and benign).

Bonus fix in Makefile: `debug` target now appends $(TLS_DEF) +
$(TLS_FLAGS) so 'make debug' actually compiles src/catalog/user.c
(needs openssl/evp.h via pkg-config).  Without this the ASan rebuild
fails before the sanitizer ever runs.

Verification:
  - `make debug && make test` under ASan/UBSan: 0 SUMMARY lines (was 14
    firings across 3 distinct sites pre-fix).  The 1 remaining test
    failure (test_influx_line.c:419 'parse throughput > 1M lines/sec')
    is purely the sanitiser's overhead — irrelevant to correctness.
  - `make && make test` release: 58/58 PASS.

+++ src/compress/pfor.c
@@ -377,7 +377,14 @@ int tsdb_pfor_encode_i64(const int64_t *in, size_t n,
         int fits_u32 = 1;
 
         for (unsigned k = 0; k < count; k++) {
-            int64_t delta = in[i + k] - prev;
+            /* Subtract via uint64_t to avoid the int64 overflow UB that
+             * UBSAN catches when adversarial inputs span the signed
+             * range (e.g. ts ≈ +7.2e18 vs prev ≈ −6.6e18 in the test
+             * suite).  Two's-complement wrap-around through unsigned
+             * arithmetic produces the same bit pattern that the
+             * subsequent zigzag encode treats as the signed delta —
+             * portable & defined. */
+            int64_t delta = (int64_t)((uint64_t)in[i + k] - (uint64_t)prev);
             uint64_t zz = tsdb_zigzag_enc(delta);
             i64_abs[k] = in[i + k];
             if (zz > 0xFFFFFFFFu) {
